requirements.txt
streamlit>=1.28
groq>=0.9.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
numpy>=1.24
python-dotenv>=1.0.0
ragas>=0.1.0
datasets>=2.14.0
